No exceptions during metric creation
Our metrics collection use case involves creating metrics dynamically
and in different scopes. We noticed that one of the prime contributors
to thrown exceptions was the `HdrMetricBuilder`. By trying to register
the metric again the error case is only catched with an exception which
does hit our performance to some degree.

In order to remove this exception based error handling this change first
checks whether the metric registry already contains the metric
explicitly. If it does, the metric is only returned at this point. If
not, it is being safely created.